+++ src/test/java/kitchenpos/fixture/MenuFixtureFactory.java
@@ -2,9 +2,9 @@ package kitchenpos.fixture;
 
 import java.math.BigDecimal;
 import java.util.List;
-import kitchenpos.domain.Menu;
+import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuGroup;
-import kitchenpos.domain.MenuProduct;
+import kitchenpos.menu.domain.MenuProduct;
 
 public class MenuFixtureFactory {
     private MenuFixtureFactory() {
