+++ src/test/java/kitchenpos/fixture/MenuProductFixtureFactory.java
@@ -1,6 +1,6 @@
 package kitchenpos.fixture;
 
-import kitchenpos.domain.MenuProduct;
+import kitchenpos.menu.domain.MenuProduct;
 
 public class MenuProductFixtureFactory {
     private MenuProductFixtureFactory() {
